# api/openapi.yml
openapi: 3.0.3
info:
  title: audited
  description: API reference for audited API
  termsOfService: https://github.com/firminochangani/audited
  contact:
    email: firmino.changani@gmail.com
  license:
    name: MIT
    url: https://github.com/firminochangani/audited/LICENSE.md
  version: 0.0.1
...
servers:
  - url: http://localhost:8080
paths:
...
  /x/v1/sources:
    get:
      summary: List all sources
      description: ...
      tags:
        - Admin API
      operationId: getSources
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/pagination_limit'
        - $ref: '#/components/parameters/pagination_page'
      responses:
        "200":
          description: A paginated list of sources
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceList"
        default:
          $ref: "#/components/responses/InternalServerError"
    post:
      summary: Create a source
      description: ...
      tags:
        - Admin API
      operationId: createSource
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required: [ name ]
              properties:
                name:
                  type: string
      responses:
        "204":
          description: Source created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Source"
        default:
          $ref: "#/components/responses/InternalServerError"
...
components:
  parameters:
...
    pagination_limit:
      name: limit
      description: The number of items per page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
    pagination_page:
      name: page
      description: The page number
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
  schemas:
...
    Source:
      type: object
      required: [id, name, created_at, updated_at]
      properties:
        id:
          type: string
          x-order: 1
        name:
          type: string
          x-order: 2
        created_at:
          type: string
          format: date-time
          x-order: 3
        updated_at:
          type: string
          format: date-time
          x-order: 4
    SourceList:
      required: [data, pagination]
      properties:
        pagination:
          $ref: '#/components/schemas/Pagination'
        data:
          type: array
          items:
            $ref: '#/components/schemas/Source'
    # Event type
...
    Pagination:
      required: [ total, per_page, current_page, total_pages ]
      properties:
        total:
          type: integer
          x-order: 1
        per_page:
          type: integer
          x-order: 2
        current_page:
          type: integer
          x-order: 3
        total_pages:
          type: integer
          x-order: 4

    # Error
    ErrorSchema:
      type: object
      required: [error, message]
      properties:
        message:
          type: string
          description: A description about the error
        error:
          type: string
          description: Error custom error code such as 'email_in_use'
  responses:
...
    InternalServerError:
      description: "A generic error response"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorSchema"
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    TokenAuth:
      type: apiKey
      in: header
      name: x-token